+++ src/ElasticSan/ElasticSan/ChangeLog.md
@@ -19,6 +19,12 @@
 -->
 ## Upcoming Release
 * Added support for soft delete on volume group and volumes
+    - `New-AzElasticSanVolumeGroup`
+    - `Get-AzElasticSanVolumeGroup`
+    - `Get-AzElasticSanVolume`
+    - `Remove-AzElasticSanVolume`
+    - `Restore-AzElasticSanVolume`
+
 * Added support for whether disk snapshot backup can be taken for volumes.
     - `Test-AzElasticSanVolumeBackup`
     - `Test-AzElasticSanVolumeRestore`